@@ -2,7 +2,7 @@ package realtime
 
 import (
 	"encoding/json"
-	"github.com/gabstv/melody"
+	"errors"
 	log "github.com/sirupsen/logrus"
 	"net/http"
 )
@@ -20,29 +20,37 @@ type Message struct {
 }
 
 type Realtime struct {
-	melody   *melody.Melody
-	clients  ClientStore
-	channels ChannelStore
+	connector *Connector
+	channels  ChannelStore
 }
 
 // New Creates a new Realtime instance
-func New() *Realtime {
-	instance := Realtime{}
-	instance.init()
-	return &instance
+func New(connector *Connector) *Realtime {
+	r := Realtime{}
+
+	r.connector = connector
+	r.channels.init()
+	r.connector.hook(&Hooks{
+		OnConnect:    r.connectHandler,
+		OnDisconnect: r.disconnectHandler,
+		OnMessage:    r.messageHandler,
+	})
+
+	return &r
 }
 
-type handlerFunc func(c *Client)
-type handlerDataFunc func(c *Client, data []byte)
-
 // RegisterChannel registers a new channel
-func (r *Realtime) RegisterChannel(channelName string, handlers ChannelHandlers) {
-	r.channels.Register(channelName, handlers)
+func (r *Realtime) RegisterChannel(channelName string, handlers ChannelHandlers) *Channel {
+	return r.channels.Register(channelName, handlers)
 }
 
 // HandleRequest handles a new websocket request, adds the properties to the new client
 func (r *Realtime) HandleRequest(writer http.ResponseWriter, request *http.Request, properties map[string]interface{}) error {
-	return r.melody.HandleRequestWithKeys(writer, request, properties)
+	return r.connector.requestHandler(writer, request, properties)
+}
+
+func (r *Realtime) IsConnected(clientId string) bool {
+	return r.connector.clients.Exists(clientId)
 }
 
 // IsSubscribed checks whether a client is subscribed to a certain channelPath or not
@@ -53,44 +61,24 @@ func (r *Realtime) IsSubscribed(channelPath string, clientId string) bool {
 	return false
 }
 
-// mapEventToClient maps a melody callback to a realtime callback with the according Client
-func (r *Realtime) mapEventToClient(handler handlerFunc) func(*melody.Session) {
-	return func(s *melody.Session) {
-		id, _ := s.Get("id")
-		client := r.clients.Get(id.(string))
-		handler(client)
+func (r *Realtime) Send(channelPath string, clientId string, payload []byte) error {
+	channelExists, channel, _ := r.channels.Get(channelPath)
+	if !channelExists {
+		return errors.New("channel does not exists")
 	}
-}
-
-// mapDataEventToClient maps a melody callback with data to a realtime callback with the according Client
-func (r *Realtime) mapDataEventToClient(handler handlerDataFunc) func(*melody.Session, []byte) {
-	return func(s *melody.Session, data []byte) {
-		id, _ := s.Get("id")
-		client := r.clients.Get(id.(string))
-		handler(client, data)
+	context, userSubscribed := channel.FindContext(clientId, channelPath)
+	if !userSubscribed {
+		return errors.New("user not subscribed to channel")
 	}
-}
-
-// init Initializes a realtime instance
-func (r *Realtime) init() {
-	r.melody = melody.New()
-	r.clients.init()
-	r.channels.init()
-	r.melody.HandleConnect(r.connectHandler)
-	r.melody.HandleDisconnect(r.mapEventToClient(r.disconnectHandler))
-	r.melody.HandleMessage(r.mapDataEventToClient(r.messageHandler))
+	return context.Send(payload)
 }
 
 // connectHandler handles a new melody connection
-func (r *Realtime) connectHandler(s *melody.Session) {
-	client := r.clients.Join(s, s.Keys())
-	s.Set("id", client.Id)
-}
+func (r *Realtime) connectHandler(client *Client) {}
 
 // disconnectHandler handles a client disconnect
 func (r *Realtime) disconnectHandler(c *Client) {
 	r.channels.UnsubscribeAll(c.Id)
-	r.clients.Remove(c.Id)
 }
 
 // messageHandler handles a new client message
